add last_time on database page
